move cache keys enum to redis module
